+++ CHANGELOG.md
@@ -62,6 +62,7 @@
 * Changed `Affine` to always use `order=3` for heatmap augmentation.
 * Fixed `PiecewiseAffine` heatmap augmentation producing arrays with values outside the range `[0.0, 1.0]` when `order` was set to `3`.
 * Changed `PiecewiseAffine` to always use `order=3` for heatmap augmentation.
+* Changed `ElasticTransformation` to always use `order=3` for heatmap augmentation.
 
 
 # 0.2.8